refactor of filter products components
