+++ src/components/product/ProductList.js
@@ -1,17 +1,21 @@
-import { useEffect, useState } from "react"
-import { useDispatch, useSelector } from "react-redux"
-import { FILTER_BY_CATEGORY, FILTER_BY_SEARCH, selectFilteredProducts, SORT_PRODUCTS } from "../../redux/slice/filterSlice"
+import { useState } from "react"
+import { useSelector } from "react-redux"
+import { selectFilteredProducts } from "../../redux/slice/filterSlice"
 import { Pagination } from "../pagination/Pagination"
-import { Search } from "./Search"
+import { Search } from "./Filters/Search"
 import { ProductItem } from "./ProductItem"
-import { Sort } from "./Sort"
+import { Sort } from "./Filters/Sort"
+
+import { CategoryFilter } from "./Filters/CategoryFilter"
+import { PriceFilter } from "./Filters/PriceFilter"
+// import { ClearAll } from "./Filters/ClearAll"
+
 
 
 export const ProductList = ({ products }) => {
   const [search, setSearch] = useState('')
-  const [sort, setSort] = useState('Latest')
-  const dispatch = useDispatch()
   const filteredProducts = useSelector(selectFilteredProducts)
+
   // Pagination states
   const [currentPage, setCurrentPage] = useState(1)
   const [productsPerPage, setProductsPerPage] = useState(4)
@@ -21,30 +25,6 @@ export const ProductList = ({ products }) => {
   const indexOfFirstProduct = indexOfLastProduct - productsPerPage
   const currentProducts = filteredProducts.slice(indexOfFirstProduct, indexOfLastProduct)
 
-  useEffect(() => {
-    dispatch(FILTER_BY_SEARCH({
-      products, search
-    }))
-  }, [search, products, dispatch])
-
-  useEffect(() => {
-    dispatch(SORT_PRODUCTS({
-      products, sort
-    }))
-  }, [sort, products, dispatch])
-
-  const allCategories = [
-    "All",
-    ...new Set(products.map((product) => product.category))
-  ]
-  const [category, setCategory] = useState('All')
-
-  const filterProducts = (cat) => {
-    setCategory(cat)
-    dispatch(FILTER_BY_CATEGORY({
-      products, category: cat
-    }))
-  }
   return (
     <>
       {/* ------Filters Navbar-------- */}
@@ -52,7 +32,6 @@ export const ProductList = ({ products }) => {
         <div className="mx-auto  px-2 sm:px-6 lg:px-8">
           <div className="relative flex h-16 items-center justify-between">
             <div className="flex flex-1 items-center justify-center sm:items-stretch sm:justify-start">
-            
               <div className="hidden sm:ml-6 sm:block">
                 <div className="flex space-x-4">
                   <Search value={search} onChange={(e) => setSearch(e.target.value)} />
@@ -60,20 +39,17 @@ export const ProductList = ({ products }) => {
               </div>
             </div>
             <div className="hidden sm:ml-6 sm:block">
-              <Sort sort={sort} setSort={setSort} className=' px-3 py-2 rounded-md text-sm font-medium' />
+              <Sort className=' px-3 py-2 rounded-md text-sm font-medium' />
             </div>
-            <div className="relative ml-3">
-              <div className="hidden sm:ml-6 sm:block">
-                <select value={category} onChange={(e) => filterProducts(e.target.value)} className='select select-primary select-sm w-full max-w-xs'>
-
-                  {allCategories.map((cat, index) => {
-                    return (
-                      <option key={index} >{cat}</option>
-                    )
-                  })}
-                </select>
-              </div>
+            <div className="hidden sm:ml-6 sm:block">
+              <CategoryFilter className=' px-3 py-2 rounded-md text-sm font-medium' />
             </div>
+            <div className="hidden sm:ml-6 sm:block">
+              <PriceFilter />
+            </div>
+            {/* <div className="hidden sm:ml-6 sm:block">
+              <ClearAll/>
+            </div> */}
           </div>
         </div>
 
@@ -82,21 +58,21 @@ export const ProductList = ({ products }) => {
           {/* <a href="/#" className="bg-gray-900 text-white block px-3 py-2 rounded-md text-base font-medium" aria-current="page">Products Filter</a> */}
           <div className=" flex flex-col items-center w-auto px-2 pt-2 pb-3 gap-2">
             {/* <!-- Current: "bg-gray-900 text-white", Default: "text-gray-300 hover:bg-gray-700 hover:text-white" --> */}
-            <Search value={search} onChange={(e) => setSearch(e.target.value)} />
-            <div className=" flex gap-2 ">
-              <Sort sort={sort} setSort={setSort} className=' px-3 py-2 rounded-md text-sm font-medium' />
-           
-              <select value={category} onChange={(e) => filterProducts(e.target.value)} className='select select-primary select-sm '>
-
-                {allCategories.map((cat, index) => {
-                  return (
-                    <option key={index} >{cat}</option>
-                  )
-                })}
-              </select>
-              </div>
+            <Search />
+            <div className="flex ">
+              <Sort className='px-3 py-2 rounded-md text-sm font-medium  ' />
             </div>
+            <div className="flex ">
+              <CategoryFilter className='px-3 py-2 rounded-md text-sm font-medium  ' />
+            </div>
+            <div className="flex ">
+              <PriceFilter />
+            </div>
+            {/* <div className="flex ">
+            <ClearAll />
+            </div> */}
           </div>
+        </div>
       </nav>
 
       {/* ------Images Section-------- */}
